be_apigateway/requirements.txt
aniso8601==9.0.1
autopep8==2.0.1
boto3==1.26.78
botocore==1.29.78
click==8.1.3
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
s3transfer==0.6.0
schedule==1.1.0
serverless-wsgi==3.0.1
six==1.16.0
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.14.0
